+++ src/hike/support/__init__.py
@@ -3,10 +3,11 @@
 ##############################################################################
 # Local imports.
 from .history import History
+from .mouse import is_copy_request_click
 from .view_in_browser import view_in_browser
 
 ##############################################################################
 # Exports.
-__all__ = ["History", "view_in_browser"]
+__all__ = ["History", "is_copy_request_click", "view_in_browser"]
 
 ### __init__.py ends here