+++ ChangeLog.md
@@ -8,8 +8,12 @@
   ([#2](https://github.com/davep/hike/pull/2))
 - The suggester for the command line now prefers newer history entries over
   older. ([#3](https://github.com/davep/hike/pull/3))
-- Add the ability to move the command line to the top of the screen.
+- Added the ability to move the command line to the top of the screen.
   ([#5](https://github.com/davep/hike/pull/5))
+- Added the ability to copy the current location to the clipboard.
+  ([#9](https://github.com/davep/hike/pull/9))
+- Added the ability to copy the currently-viewed Markdown to the clipboard.
+  ([#9](https://github.com/davep/hike/pull/9))
 
 ## v0.1.0